requirements.txt
filterpy==1.4.5
lap==0.4.0
matplotlib==3.6.2
numpy==1.23.4
opencv_python==4.6.0.66
Pillow==9.3.0
scikit_image==0.19.3
scikit_learn==1.1.3
scipy==1.9.3
skimage==0.0
torch==1.13.0
tqdm==4.64.1
